requirements.txt
asgiref==3.5.2
Django==4.0.4
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.1.1
PyJWT==2.4.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
